+++ src/main/java/cl/duoc/microservicioTienda/repository/TiendaRespository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import cl.duoc.microservicioTienda.model.Tienda;
 
-public interface TiendaRespository extends JpaRepository<Tienda, Long>{
+public interface TiendaRespository extends JpaRepository<Tienda, Integer>{
 
 }
